+++ Alligator/Commands/TabItemOrders/DeleteOrderWindowOfOrderInfoCommand.cs
@@ -27,13 +27,18 @@ namespace Alligator.UI.Commands.TabItemOrders
 
         public override void Execute(object parameter)
         {
-            _orderReviewService.DeleteOrderReviewModelByOrderId(_viewModel.SelectedOrder.Id);
-            _orderDetailService.DeleteOrderDetailModelByOrderId(_viewModel.SelectedOrder.Id);
-            _orderService.DeleteOrderModel(_viewModel.SelectedOrder.Id);
-            _viewModel.AllOrders.Remove(_viewModel.SelectedOrder);
-            _viewModel.AddOrderWindowVisibility = Visibility.Collapsed;
-            _viewModel.OrdersInfoWindowVisibility = Visibility.Collapsed;
-            _viewModel.OrdersWindowVisibility = Visibility.Visible;
+
+            var userAnswer = MessageBox.Show("Удалить заказ?", "Удаление", MessageBoxButton.YesNo, MessageBoxImage.Question);
+            if (userAnswer == MessageBoxResult.Yes)
+            {
+                _orderReviewService.DeleteOrderReviewModelByOrderId(_viewModel.SelectedOrder.Id);
+                _orderDetailService.DeleteOrderDetailModelByOrderId(_viewModel.SelectedOrder.Id);
+                _orderService.DeleteOrderModel(_viewModel.SelectedOrder.Id);
+                _viewModel.AllOrders.Remove(_viewModel.SelectedOrder);
+                _viewModel.AddOrderWindowVisibility = Visibility.Collapsed;
+                _viewModel.OrdersInfoWindowVisibility = Visibility.Collapsed;
+                _viewModel.OrdersWindowVisibility = Visibility.Visible;
+            }
         }
     
     }
